Removes the `en_US` locale
Removes the `en_US` locale and updates the language selection in the design constructor.

The `en_US` locale was redundant as it mirrored the `en` locale. The language selection component is updated to use a value prop to maintain consistency and allow for default "App Setup (or English)" option, which maps to an empty string internally. Updates the constructor store to use an empty string instead of undefined to indicate no language is selected.