Adds swap functionality to PayPass
Adds the ability to include swap information in the PayPass flow.

This change introduces the "swap" parameter to the PayPass configuration,
allowing users to specify a swap value that will be displayed on the pass.

Also refactors the `getLink` helper to accept a `designData` object
instead of a boolean flag for design, providing more flexibility for
future design-related customizations.