Adds CSRF debug information
Improves the debugging experience for CSRF protection issues
by providing more detailed information when a CSRF error is
encountered at the /pass endpoint with debug mode enabled.

This includes content type, origin, base origin and whether or
not the origin matches to help the user understand the reason for
the CSRF failure. It also includes a note for how to solve cross-
origin requests.